zk-cli/requirements.txt
# CLI Framework
typer>=0.12.0
rich>=13.0.0

# AI/Embeddings
sentence-transformers>=3.0

# Vector Database
chromadb>=0.5.0

# Knowledge Graph
networkx>=3.0

# File Monitoring
watchdog>=3.0

# Data Processing
pyyaml>=6.0
pydantic>=2.0

# Template Engine
jinja2>=3.1.0
